week-4/voice_bot/requirements.txt
groq
python-dotenv
streamlit
soundfile
